+++ collector-exoplayer-example/src/main/java/com/bitmovin/analytics/exoplayer/example/MainActivity.java
@@ -11,6 +11,7 @@ import androidx.appcompat.app.AppCompatActivity;
 import com.bitmovin.analytics.BitmovinAnalytics;
 import com.bitmovin.analytics.BitmovinAnalyticsConfig;
 import com.bitmovin.analytics.data.AdEventData;
+import com.bitmovin.analytics.data.CustomData;
 import com.bitmovin.analytics.data.EventData;
 import com.bitmovin.analytics.enums.CDNProvider;
 import com.bitmovin.analytics.exoplayer.ExoPlayerCollector;
@@ -41,6 +42,7 @@ public class MainActivity extends AppCompatActivity
     private Button releaseButton;
     private Button createButton;
     private Button sourceChangeButton;
+    private Button setCustomDataButton;
     private TextView eventLogView;
 
     private DataSource.Factory dataSourceFactory;
@@ -60,6 +62,8 @@ public class MainActivity extends AppCompatActivity
         createButton.setOnClickListener(this);
         sourceChangeButton = findViewById(R.id.source_change_button);
         sourceChangeButton.setOnClickListener(this);
+        setCustomDataButton = findViewById(R.id.set_custom_data);
+        setCustomDataButton.setOnClickListener(this);
         eventLogView = findViewById(R.id.eventLog);
 
         dataSourceFactory = new DefaultDataSourceFactory(this, buildHttpDataSourceFactory());
@@ -117,7 +121,7 @@ public class MainActivity extends AppCompatActivity
             bitmovinAnalyticsConfig.setPath("/vod/new/");
             bitmovinAnalyticsConfig.setHeartbeatInterval(59700);
             bitmovinAnalyticsConfig.setIsLive(false);
-            bitmovinAnalyticsConfig.getConfig().setTryResendDataOnFailedConnection(true);
+            bitmovinAnalyticsConfig.getConfig().setTryResendDataOnFailedConnection(false);
 
             eventLogView.setText("");
 
@@ -231,6 +235,8 @@ public class MainActivity extends AppCompatActivity
             createPlayer();
         } else if (v == sourceChangeButton) {
             changeSource();
+        } else if (v == setCustomDataButton) {
+            setCustomData();
         }
     }
 
@@ -243,6 +249,13 @@ public class MainActivity extends AppCompatActivity
 
         bitmovinAnalytics.attachPlayer(player);
         player.setMediaSource(dashMediaSource);
+    }
+
+    private void setCustomData() {
+        CustomData customData = bitmovinAnalytics.getCustomData();
+        customData.setCustomData1("custom_data_1_changed");
+        customData.setCustomData2("custom_data_2_changed");
+        bitmovinAnalytics.setCustomData(customData);
     }
 
     @Override